Remove settings helper class
