+++ src/Appwrite/Platform/Modules/Project/Http/Project/Auth/Ldap/Update.php
@@ -3,7 +3,6 @@
 namespace Appwrite\Platform\Modules\Project\Http\Project\Auth\Ldap;
 
 use Appwrite\Auth\LDAP\Client;
-use Appwrite\Auth\LDAP\Settings;
 use Appwrite\Event\Event;
 use Appwrite\Extend\Exception;
 use Appwrite\Platform\Action;
@@ -62,12 +61,12 @@ class Update extends Action
             ))
             ->param('host', null, new Nullable(new Text(256, 0)), 'Directory hostname or IP address.', optional: true)
             ->param('port', null, new Nullable(new Integer()), 'Directory port. Defaults to 389, or 636 with SSL.', optional: true)
-            ->param('encryption', null, new Nullable(new WhiteList(Settings::ENCRYPTIONS, true)), 'Transport security. A simple bind sends the password in the clear, so "none" should only be used on a trusted network.', optional: true, enum: new Enum(name: 'ProjectAuthLdapEncryption'))
+            ->param('encryption', null, new Nullable(new WhiteList(Client::ENCRYPTIONS, true)), 'Transport security. A simple bind sends the password in the clear, so "none" should only be used on a trusted network.', optional: true, enum: new Enum(name: 'ProjectAuthLdapEncryption'))
             ->param('baseDn', null, new Nullable(new Text(512, 0)), 'Subtree the user search starts from. For example: ou=people,dc=example,dc=com', optional: true)
             ->param('bindDn', null, new Nullable(new Text(512, 0)), 'Service account used to search for users. Leave empty if the directory allows anonymous search.', optional: true)
             ->param('bindPassword', null, new Nullable(new Text(512, 0)), 'Service account password.', optional: true)
-            ->param('userFilter', null, new Nullable(new Text(1024, 0)), 'Search filter locating the user, containing the ' . Settings::PLACEHOLDER . ' placeholder. For example: (uid=' . Settings::PLACEHOLDER . ')', optional: true)
-            ->param('provisionFilter', null, new Nullable(new Text(1024, 0)), 'Optional filter a user must also match to be allowed an account, typically a group membership. ' . Settings::PLACEHOLDER . ' is replaced with the entry\'s distinguished name, so a group check reads (&(cn=staff)(member=' . Settings::PLACEHOLDER . ')). The match must name the entry signing in. Evaluated on every sign-in, so removing someone from the group revokes their access.', optional: true)
+            ->param('userFilter', null, new Nullable(new Text(1024, 0)), 'Search filter locating the user, containing the ' . Client::PLACEHOLDER . ' placeholder. For example: (uid=' . Client::PLACEHOLDER . ')', optional: true)
+            ->param('provisionFilter', null, new Nullable(new Text(1024, 0)), 'Optional filter a user must also match to be allowed an account, typically a group membership. ' . Client::PLACEHOLDER . ' is replaced with the entry\'s distinguished name, so a group check reads (&(cn=staff)(member=' . Client::PLACEHOLDER . ')). The match must name the entry signing in. Evaluated on every sign-in, so removing someone from the group revokes their access.', optional: true)
             ->param('emailAttribute', null, new Nullable(new Text(128, 0)), 'Attribute holding the email address. Required, because an account cannot be created without one.', optional: true)
             ->param('nameAttribute', null, new Nullable(new Text(128, 0)), 'Attribute holding the display name.', optional: true)
             ->param('enabled', null, new Nullable(new Boolean()), 'LDAP sign-in status. Setting this to true validates the configuration and throws if the directory cannot be reached.', optional: true)
@@ -106,15 +105,15 @@ class Update extends Action
 
         // TODO: Fix when adding array support
         $directory = $directories[0] ?? [];
-        
+
         $directory = [
             'host' => $host ?? ($directory['host'] ?? ''),
-            'port' => $port ?? ($directory['port'] ?? Settings::DEFAULT_PORT),
-            'encryption' => $encryption ?? ($directory['encryption'] ?? Settings::ENCRYPTION_TLS),
+            'port' => $port ?? ($directory['port'] ?? Client::DEFAULT_PORT),
+            'encryption' => $encryption ?? ($directory['encryption'] ?? Client::ENCRYPTION_TLS),
             'baseDn' => $baseDn ?? ($directory['baseDn'] ?? ''),
             'bindDn' => $bindDn ?? ($directory['bindDn'] ?? ''),
             'bindPassword' => $bindPassword ?? ($directory['bindPassword'] ?? ''),
-            'userFilter' => $userFilter ?? ($directory['userFilter'] ?? '(uid=' . Settings::PLACEHOLDER . ')'),
+            'userFilter' => $userFilter ?? ($directory['userFilter'] ?? '(uid=' . Client::PLACEHOLDER . ')'),
             'provisionFilter' => $provisionFilter ?? ($directory['provisionFilter'] ?? ''),
             'emailAttribute' => $emailAttribute ?? ($directory['emailAttribute'] ?? 'mail'),
             'nameAttribute' => $nameAttribute ?? ($directory['nameAttribute'] ?? 'cn'),
@@ -122,7 +121,7 @@ class Update extends Action
 
         if ($enabled === true) {
             try {
-                $settings = new Settings(
+                $client = new Client(
                     host: $directory['host'],
                     port: (int)$directory['port'],
                     encryption: $directory['encryption'],
@@ -134,8 +133,8 @@ class Update extends Action
                     emailAttribute: $directory['emailAttribute'],
                     nameAttribute: $directory['nameAttribute'],
                 );
-    
-                (new Client($settings))->verify();
+
+                $client->verify();
             } catch (\Throwable $error) {
                 throw new Exception(Exception::GENERAL_ARGUMENT_INVALID, 'Could not enable LDAP: ' . $error->getMessage());
             }
@@ -157,15 +156,15 @@ class Update extends Action
         $response->dynamic(new Document([
             '$id' => 'ldap',
             'enabled' => $auths[Config::getParam('auth')['ldap']['key']] ?? false,
-            'host' => $directory['host'] ?? '',
-            'port' => (int)($directory['port'] ?? Settings::DEFAULT_PORT),
-            'encryption' => $directory['encryption'] ?? Settings::ENCRYPTION_TLS,
-            'baseDn' => $directory['baseDn'] ?? '',
-            'bindDn' => $directory['bindDn'] ?? '',
-            'userFilter' => $directory['userFilter'] ?? '',
-            'provisionFilter' => $directory['provisionFilter'] ?? '',
-            'emailAttribute' => $directory['emailAttribute'] ?? 'mail',
-            'nameAttribute' => $directory['nameAttribute'] ?? 'cn',
+            'host' => $directory['host'],
+            'port' => (int)$directory['port'],
+            'encryption' => $directory['encryption'],
+            'baseDn' => $directory['baseDn'],
+            'bindDn' => $directory['bindDn'],
+            'userFilter' => $directory['userFilter'],
+            'provisionFilter' => $directory['provisionFilter'],
+            'emailAttribute' => $directory['emailAttribute'],
+            'nameAttribute' => $directory['nameAttribute'],
         ]), Response::MODEL_AUTH_LDAP);
     }
 }
